+++ lib/src/query/enum_query_type.dart
@@ -6,6 +6,7 @@ enum EnumQueryType {
   update,
   updateOne,
   delete,
+  deleteOne,
   search,
   getAll,
   conformToTemplate, // DB shape change.