+++ lib/src/query/query_result.dart
@@ -1,27 +1,34 @@
-import 'package:file_state_manager/file_state_manager.dart';
-
 import '../../delta_trace_db.dart';
 
+/// (en) This class stores the query results and additional information from
+/// the database.
+///
 /// (ja) DBへのクエリ結果や付加情報を格納したクラスです。
-class QueryResult<T> extends CloneableFile {
+class QueryResult<T> extends QueryExecutionResult {
   static const String className = "QueryResult";
-  static const String version = "1";
-  bool isNoErrors;
+  static const String version = "2";
   List<Map<String, dynamic>> result;
   int dbLength;
   int updateCount;
   int hitCount;
   String? errorMessage;
 
-  /// * [isNoErrors] : 操作が成功したかどうかのフラグ。
-  /// これはエラーが発生していないかどうかを表すため、検索や更新数が0でもtrueになります。
-  /// * [result] : 検索結果、更新結果、削除されたオブジェクトなど。
-  /// * [dbLength] : DB内のデータの総数。
-  /// * [updateCount] : 更新、または削除されたデータの総数。
-  /// * [hitCount] : 検索対象になったデータの総数。
-  /// * [errorMessage] : エラー発生時のみ追加されるメッセージ。
+  /// * [isNoErrors] : A flag indicating whether the operation was successful.
+  /// This also changes depending on the value of the optional argument
+  /// mustAffectAtLeastOne when creating a query.
+  /// When mustAffectAtLeastOne is true,
+  /// if the number of operation targets is 0,
+  /// it will be treated as an error and the value will be false.
+  /// When false, the value will be true even if the number of updates is 0.
+  /// In other cases, if an exception occurs internally,
+  /// it will be treated as an error.
+  /// * [result] : Search results, update results, deleted objects, etc.
+  /// * [dbLength] : The total number of records in the DB.
+  /// * [updateCount] : The total number of records add, updated or deleted.
+  /// * [hitCount] : The total number of records searched.
+  /// * [errorMessage] : A message that is added only if an error occurs.
   QueryResult({
-    required this.isNoErrors,
+    required super.isNoErrors,
     required this.result,
     required this.dbLength,
     required this.updateCount,
@@ -45,9 +52,14 @@ class QueryResult<T> extends CloneableFile {
     );
   }
 
-  /// 検索結果を指定クラスの配列で取得します。
-  /// * [fromDict] : クラスTがCloneableFileである場合、
-  /// 辞書からオブジェクトを復元するためのfromDict相当の関数を渡します。
+  /// (en) The search results will be retrieved as an array of
+  /// the specified class.
+  ///
+  /// (ja) 検索結果を指定クラスの配列で取得します。
+  ///
+  /// * [fromDict] : If class T is CloneableFile,
+  /// Pass a function equivalent to fromDict to restore the object from
+  /// a dictionary.
   List<T> convert(T Function(Map<String, dynamic>) fromDict) {
     List<T> r = [];
     for (Map<String, dynamic> i in result) {